Break out flood fill into its own package
