+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod bindings;
 pub mod convert;
 pub mod dealias_region;
 pub mod filter;
+pub mod flood_fill;
 pub mod model;
 pub mod pymodel;
 pub mod util;
